pyproject.toml
[tool.poetry]
name = "task-tracker"
version = "0.1.0"
description = ""
authors = ["artem_b <hhugoo11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
djangorestframework = "^3.9"
django-filter = "^2.1"
coreapi = "^2.3"

[tool.poetry.dev-dependencies]
pytest = "^4.2"
pytest-django = "^3.4"
pytest-cov = "^2.6"
pytest-testmon = "^0.9.14"
factory-boy = "^2.11"
mypy = "^0.660.0"
ipython = "^7.2"
flake8 = "^3.7"
pytest-factoryboy = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
